Fix bug where Jump button would have incorrect state.
If the user switched to another fragment then the savedInstanceState
would be null and so no check would be made for carrying out the state change.
Since we now have background checks, we can see if the LocationService is running
and infer from that what the jump button state should be.